@@ -1,18 +1,19 @@
-'use strict';
-
-const enableDestroy = require('server-destroy');
-const selfsigned = require('selfsigned');
-const { URL } = require('url');
-const https = require('https');
+import enableDestroy from 'server-destroy';
+import selfsigned from 'selfsigned';
+import { URL } from 'url';
+import https from 'https';
 
 const attrs = [{ name: 'commonName', value: 'podium-lib.io' }];
 const PEMS = selfsigned.generate(attrs, { days: 365 });
 
-class HttpsServer {
+export class HttpsServer {
+    address = '';
+    /** @type {import('http').RequestListener | undefined} */
+    request = undefined;
+    /** @type {import('https').Server | undefined} */
+    server = undefined;
+
     constructor() {
-        this.address = '';
-        this.request = undefined;
-        this.server = undefined;
         this.app = https.createServer(
             {
                 key: PEMS.private,
@@ -33,9 +34,10 @@ class HttpsServer {
     listen() {
         return new Promise((resolve) => {
             this.server = this.app.listen(0, '0.0.0.0', () => {
-                this.address = `https://${this.server.address().address}:${
-                    this.server.address().port
-                }`;
+                let address = /** @type {import('net').AddressInfo} */ (
+                    this.server.address()
+                );
+                this.address = `https://${address.address}:${address.port}`;
                 resolve(this.address);
             });
             enableDestroy(this.server);
@@ -94,5 +96,3 @@ class HttpsServer {
         });
     }
 }
-
-module.exports = HttpsServer;
