@@ -1,9 +1,22 @@
-'use strict';
+import { URL } from 'url';
+import http from 'http';
 
-const { URL } = require('url');
-const http = require('http');
+/**
+ * @typedef {object} RequestOptions
+ * @property {string} [pathname="/"]
+ * @property {string} [address=""]
+ * @property {Record<string, string>} [headers={}]
+ * @property {string} [method="GET"]
+ * @property {boolean} [json=false]
+ */
 
-const request = (
+/**
+ * @template T
+ * @param {RequestOptions} options
+ * @param {T} payload
+ * @returns {Promise<{ headers: import('http').IncomingHttpHeaders, body: T}>}
+ */
+export const request = (
     {
         pathname = '/',
         address = '',
@@ -19,6 +32,7 @@ const request = (
         if (method === 'POST' || method === 'PUT' || method === 'DELETE') {
             headers = Object.assign(headers, {
                 'Content-Type': 'application/x-www-form-urlencoded',
+                // @ts-expect-error It's fine
                 'Content-Length': Buffer.byteLength(payload),
             });
         }
@@ -60,4 +74,3 @@ const request = (
 
         req.end();
     });
-module.exports = request;
