@@ -1,14 +1,15 @@
-'use strict';
+import enableDestroy from 'server-destroy';
+import http from 'http';
+import url from 'url';
 
-const enableDestroy = require('server-destroy');
-const http = require('http');
-const url = require('url');
+export class HttpServer {
+    address = '';
+    /** @type {import('http').RequestListener | undefined} */
+    request = undefined;
+    /** @type {import('http').Server | undefined} */
+    server = undefined;
 
-class HttpServer {
     constructor() {
-        this.address = '';
-        this.request = undefined;
-        this.server = undefined;
         this.app = http.createServer((req, res) => {
             if (this.request) {
                 this.request(req, res);
@@ -23,9 +24,10 @@ class HttpServer {
     listen() {
         return new Promise((resolve) => {
             this.server = this.app.listen(0, '0.0.0.0', () => {
-                this.address = `http://${this.server.address().address}:${
-                    this.server.address().port
-                }`;
+                let address = /** @type {import('net').AddressInfo} */ (
+                    this.server.address()
+                );
+                this.address = `http://${address.address}:${address.port}`;
                 resolve(this.address);
             });
             enableDestroy(this.server);
@@ -75,5 +77,3 @@ class HttpServer {
         });
     }
 }
-
-module.exports = HttpServer;
