@@ -1,14 +1,17 @@
-'use strict';
+import { PodletServer } from './podlet-server/server.js';
+import { HttpsServer } from './https-server/server.js';
+import { HttpServer } from './http-server/server.js';
+import { request } from './http-request/request.js';
+import {
+    destinationBufferStream,
+    destinationObjectStream,
+} from './stream-utils/streams.js';
 
-const PodletServer = require('./podlet-server/server');
-const HttpsServer = require('./https-server/server');
-const HttpServer = require('./http-server/server');
-const request = require('./http-request/request');
-const streams = require('./stream-utils/streams');
-
-module.exports.destinationObjectStream = streams.destinationObjectStream;
-module.exports.destinationBufferStream = streams.destinationBufferStream;
-module.exports.PodletServer = PodletServer;
-module.exports.HttpsServer = HttpsServer;
-module.exports.HttpServer = HttpServer;
-module.exports.request = request;
+export {
+    PodletServer,
+    HttpsServer,
+    HttpServer,
+    request,
+    destinationBufferStream,
+    destinationObjectStream,
+};
