@@ -1,16 +1,27 @@
-/* eslint-disable no-plusplus */
-/* eslint-disable no-underscore-dangle */
+import { HttpIncoming } from '@podium/utils';
+import Podlet from '@podium/podlet';
+import enableDestroy from 'server-destroy';
+import EventEmitter from 'events';
+import http from 'http';
+import url from 'url';
 
-'use strict';
+/**
+ * @typedef {object} PodletServerOptions
+ * @property {string} [manifest="/manifest.json"]
+ * @property {string} [fallback="/fallback.html"]
+ * @property {string} [content="/index.html"]
+ * @property {string} [version="1.0.0"]
+ * @property {string} [pathname="/"]
+ * @property {string} [name="component"]
+ * @property {{ js?: string; css?: string; }} [assets]
+ * @property {Record<string, string>} [proxy]
+ */
 
-const { HttpIncoming } = require('@podium/utils');
-const enableDestroy = require('server-destroy');
-const EventEmitter = require('events');
-const Podlet = require('@podium/podlet');
-const http = require('http');
-const url = require('url');
-
-class PodletServer extends EventEmitter {
+export class PodletServer extends EventEmitter {
+    /**
+     * @constructor
+     * @param {PodletServerOptions} options
+     */
     constructor({
         manifest = '/manifest.json',
         fallback = '/fallback.html',
@@ -25,6 +36,7 @@ class PodletServer extends EventEmitter {
 
         // Private
         this._server = undefined;
+        this._manifest = {};
 
         this._podlet = new Podlet({
             manifest,
@@ -121,8 +133,8 @@ class PodletServer extends EventEmitter {
             get: () =>
                 // TODO: remove workaround
                 ({
-                    css: this._podlet.css(),
-                    js: this._podlet.js(),
+                    css: this._podlet.css([]),
+                    js: this._podlet.js([]),
                 }),
             set: (value) => {
                 // TODO: does probably not work as is (not in use since no tests break)
@@ -381,5 +393,3 @@ class PodletServer extends EventEmitter {
         return `1.0.0-beta.${index.toString()}`;
     }
 }
-
-module.exports = PodletServer;
