@@ -1,12 +1,21 @@
-'use strict';
+import stream from 'readable-stream';
 
-const stream = require('readable-stream');
-
-const destinationObjectStream = (done) => {
+/**
+ * @template [T=string]
+ * @param {(result: T[]) => void} done
+ * @returns
+ */
+export const destinationObjectStream = (done) => {
+    /** @type {T[]} */
     const arr = [];
 
     const dStream = new stream.Writable({
         objectMode: true,
+        /**
+         * @param {T} chunk
+         * @param {string} encoding
+         * @param {() => void} callback
+         */
         write(chunk, encoding, callback) {
             arr.push(chunk);
             callback();
@@ -19,13 +28,22 @@ const destinationObjectStream = (done) => {
 
     return dStream;
 };
-module.exports.destinationObjectStream = destinationObjectStream;
 
-const destinationBufferStream = (done) => {
+/**
+ * @param {(result: string) => void} done
+ * @returns
+ */
+export const destinationBufferStream = (done) => {
+    /** @type {string[]} */
     const buffer = [];
 
     const dStream = new stream.Writable({
         objectMode: false,
+        /**
+         * @param {string} chunk
+         * @param {string} encoding
+         * @param {() => void} callback
+         */
         write(chunk, encoding, callback) {
             buffer.push(chunk);
             callback();
@@ -38,4 +56,3 @@ const destinationBufferStream = (done) => {
 
     return dStream;
 };
-module.exports.destinationBufferStream = destinationBufferStream;
